+++ MiamIOSFramework/MiamIOSFramework/ui/state/ManagementResourceState.swift
@@ -46,7 +46,11 @@ struct ManagementResourceState<T : AnyObject, SuccessView : View, LoadingView : 
         case is BasicUiStateEmpty:
             emptyView
         case is BasicUiStateError:
-            HStack{} // TODO handle error state
+            if let error = Template.sharedInstance.ErrorTemplate {
+                error()
+            } else {
+                HStack{} // TODO handle error state
+            }
         case is BasicUiStateSuccess<T> :
             successView
         case is BasicUiStateLoading: